Refactor `PullRequestActor` a bit (#3371)
